requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
crewai
crewai-tools
langchain-openai
pypdf